+++ lefthk/src/config/key.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone)]
+pub enum Key {
+    Key(String),
+    Keys(Vec<String>),
+}
+
